+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # MapTiler 3D Models Changelog
 
+### NEXT
+## ✨ New Features
+- Adds `Item3D.intersects(item: Item3D, precision: "low" | "medium")` to test for intersections of Item3D objects and their associated meshes.
+- Moves clongin logic to `Item3D`, adding `Item3D.clone()` method for easier cloning of 3D items.
+
+## ⚙️ Other
+None
+
+## 🐞 Bug Fixes
+None
+
 ### 3.1.0
 
 ## ✨ New Features